refactor: Rename mamba_attn which is for mamba2 to mamba2_attn
